Add Datahub datasets various changes
The connection to datahub is established.
Datasets with the following properties will be available in gerbil:
* nif tagged
* end with .ttl
* ends NOT with dataid.ttl
* less than 20Mb

Currently the Datahub loader is no static instance.
The first time the website will be loaded the dataset names and
path to the files will be loaded.
Every time the experiment starts the information would be loaded.
This should be fixed if the 'springify' (ref #22) is done.

Removed the background loading of the tomcat (in start.sh the &).
So the tomcat is easier to stop.

Add a file logging. All log messages will now be add to the file
gerbil.log (should be set to .gitignore)

Currently there is a bug in the filepath where the nif datasets should
be written to. The path would not load properly from the
GerbilConfiguration and I did not found the problem. I will open an
issue for that.

fix #10