New experiment type KE(Knowledge extraction)
Have given the type as TypedNamedEntity as the functionality of this experiment type is to identify the named entities and linking them with the URI along with the relations among the entities. Have extended this relation only with OKETask1 and RE as OKETask1 covers all the main functionality and RE covers the relation identification.